--- .changeset/witty-icons-know.md
@@ -1,5 +1,0 @@
----
-'nextjs-website': minor
----
-
-Implement EC 'Hero' component inside Next.JS
