+++ apps/nextjs-website/package.json
@@ -23,14 +23,12 @@
     "marked": "^11.1.0",
     "next": "14.1.0",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-swipeable-views": "^0.14.0"
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
     "@types/node": "20.5.9",
     "@types/react": "18.2.21",
     "@types/react-dom": "18.2.7",
-    "@types/react-swipeable-views": "^0.13.5",
     "@typescript-eslint/eslint-plugin": "^5.59.0",
     "@vitest/coverage-v8": "^0.34.6",
     "eslint": "^8.49.0",
