+++ apps/nextjs-website/react-components/components/Accordion/AccordionItem.tsx
@@ -4,15 +4,18 @@ import AccordionSummary from '@mui/material/AccordionSummary';
 import AccordionDetails from '@mui/material/AccordionDetails';
 import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
 import { Typography } from '@mui/material';
-import { AccordionItemProps } from '../../utils/Components.types';
-import { RenderGenericBody } from '../../utils/Components.helpers';
+import { AccordionItemProps } from '../../types/Accordion/Accordion.types';
+import { Body } from '../common/Common';
+import { TextColor } from '../common/Common.helpers';
 
 export const AccordionItem: React.FC<AccordionItemProps> = ({
   header,
   content,
+  theme,
 }) => {
   const controlsId = React.useId() + '-controls';
   const headerId = React.useId() + '-header';
+  const textColor = TextColor(theme);
 
   return (
     <MUIAccordion
@@ -34,7 +37,7 @@ export const AccordionItem: React.FC<AccordionItemProps> = ({
         </Typography>
       </AccordionSummary>
       <AccordionDetails>
-        <RenderGenericBody variant='body2' body={content} />
+        <Body variant='body2' body={content} textColor={textColor} />
       </AccordionDetails>
     </MUIAccordion>
   );
