+++ apps/nextjs-website/react-components/components/index.ts
@@ -1,5 +1,6 @@
 import Hero from './Hero/Hero';
 import Editorial from './Editorial/Editorial';
 import HowTo from './HowTo/HowTo';
+import Accordion from './Accordion/Accordion';
 
-export { Hero, Editorial, HowTo };
+export { Hero, Editorial, HowTo, Accordion };
