--- apps/nextjs-website/react-components/src/index.ts
@@ -1,6 +1,0 @@
-import Hero from './components/Hero/Hero';
-import Editorial from './components/Editorial/Editorial';
-import HowTo from './components/HowTo/HowTo';
-import Accordion from './components/Accordion/Accordion';
-
-export { Hero, Editorial, HowTo, Accordion };
