+++ apps/nextjs-website/react-components/types/Accordion/Accordion.types.ts
@@ -1,0 +1,14 @@
+export interface AccordionItemProps {
+  header: string;
+  content: string | JSX.Element;
+  theme: 'light' | 'dark';
+}
+
+export interface AccordionProps {
+  title: string;
+  subtitle?: string;
+  description?: string | JSX.Element;
+  accordionItems: AccordionItemProps[];
+  theme: 'light' | 'dark';
+  layout?: 'left' | 'center' | 'right';
+}
