+++ apps/nextjs-website/react-components/types/index.ts
@@ -1,5 +1,6 @@
 import { HeroProps } from './Hero/Hero.types';
 import { EditorialProps } from './Editorial/Editorial.types';
 import { HowToProps } from './HowTo/HowTo.types';
+import { AccordionProps } from './Accordion/Accordion.types';
 
-export type { HeroProps, EditorialProps, HowToProps };
+export type { HeroProps, EditorialProps, HowToProps, AccordionProps };
