+++ apps/nextjs-website/src/components/Accordion.tsx
@@ -1,16 +1,18 @@
 'use client';
 import React from 'react';
 import MarkdownRenderer from './MarkdownRenderer';
-import { Accordion as AccordionEC } from '@react-components';
-import { AccordionProps } from '@react-components-props';
+import { Accordion as AccordionEC } from '@react-components/components';
+import { AccordionProps } from '@react-components/types';
 import { AccordionSection } from '@/lib/fetch/types/PageSection';
 
 const makeAccordionProps = ({
   subtitle,
   description,
   accordionItems,
+  theme,
   ...rest
 }: AccordionSection): AccordionProps => ({
+  theme,
   ...(subtitle && { subtitle }),
   ...(description && {
     description: MarkdownRenderer({ markdown: description, variant: 'body2' }),
@@ -18,6 +20,7 @@ const makeAccordionProps = ({
   accordionItems: accordionItems.map(({ header, content }) => ({
     header,
     content: MarkdownRenderer({ markdown: content, variant: 'body2' }),
+    theme,
   })),
   ...rest,
 });
